--- a/packages/less/src/less/tree/atrule-syntax.js
+++ b/packages/less/src/less/tree/atrule-syntax.js
@@ -1,0 +1,7 @@
+export const MediaSyntaxOptions = {
+    queryInParens: false
+};
+
+export const ContainerSyntaxOptions = {
+    queryInParens: true
+};