--- a/packages/less/src/less/tree/index.js
+++ b/packages/less/src/less/tree/index.js
@@ -25,8 +25,10 @@ import Value from './value';
 import JavaScript from './javascript';
 import Assignment from './assignment';
 import Condition from './condition';
+import QueryInParens from './query-in-parens';
 import Paren from './paren';
 import Media from './media';
+import Container from './container';
 import UnicodeDescriptor from './unicode-descriptor';
 import Negative from './negative';
 import Extend from './extend';
@@ -43,8 +45,9 @@ export default {
     Ruleset, Element, Attribute, Combinator, Selector,
     Quoted, Expression, Declaration, Call, URL, Import,
     Comment, Anonymous, Value, JavaScript, Assignment,
-    Condition, Paren, Media, UnicodeDescriptor, Negative,
-    Extend, VariableCall, NamespaceValue,
+    Condition, Paren, Media, Container, QueryInParens, 
+    UnicodeDescriptor, Negative, Extend, VariableCall, 
+    NamespaceValue,
     mixin: {
         Call: MixinCall,
         Definition: MixinDefinition